remapping script.js
previous script.js format was not corresponding with Q.10
all of the variables were placed within the local scope e.g console.log and number inputs. because all of the variables were inside the function block, it becomes executed (as a whole) when called. this caused problems in Q.10 as this code required it to reach for previous functions. when Q.10 was run, it replaced and overrided the other answers with the number input in Q.10.

by only keeping essential objects within the global scope and moving console.log and number inputs outside to the global scope made Q.10 function properly. this way all the other questions console.logs is not re-running, leaving it out of the local function.